Remove FrameSig
The intention behind FrameSig was that some signals only need to be
updated once per batch of samples rather than once per audio sample.
This mostly corresponds to user input. This caused some problems with
compatibility between components. FrameSigs could be converted to Sigs
but not the other way around, and some code was written to expect
FrameSigs. Another issue is that when a boolean value is represented by
a FrameSig, it can only chance at a cadence of once per frame, which
means that trigger signals will remain high for an entire batch of
samples rather than a single sample as was likely the intention.